refactor: organize all code
